Feature/dpr2 334 user selected columns (#40)
**User Selected Columns** 

https://dsdmoj.atlassian.net/jira/software/c/projects/DPR2/boards/1282?selectedIssue=DPR2-334

- Column select accordion
- Column options
- Apply columns functionality
- Reset columns functionality
- Reset filters while maintaining selected columns
- loading feedback
- Integration tests

**Acceptance Criteria Scenario 2 - mandatory fields:** 

Currently the API, or Field data does not include a attribute for
mandatory fields. However this PR includes the functionality to handle
disabled checkboxes and adjust the columns accordingly. A small change
is required to add the `disabled` attribute to the column data once this
it coming from the report specification