Dpr2 432 set loading triggers (#59)
- Added loading animation to "No results from query" message 

This was mentioned in the feedback where there was no loading anim
unless you opened the accordion. I suspect the state was that there were
no results from query, and they then clicked the remove filter button.
So no spinners would be shown and would give minimal feedback to the
user. So adding the loading anim here will help

- Updated loading triggers to include disabling the details accordions

Accordions are now disabled when the loading.