fix js by removing search text highligthing (#227)
Filtered report rows after searching was killing JS for those rows. Eg,
show more + bookmarking do not work on rows after performing a search.

This issue is caused by replacing the HTML in the table cells in the
search functionality to highlight text, but why this causes it to fail,
I don't know.

The PR suggests sacrificing the search term highlighting to get the JS
working again. And a further investigation can be done to fix test
highlighting without breaking the JS