Dpr2 routes refactor (#441)
### Routes refactor

Refactored route structure to better represent how the app/journeys are
put together.

Routes are split into multiple journeys:

- Download Report
- My Reports
- View Report - sync
- Request Missing Report

Still todo:
- Request report
- View Report - async

This is a bit more work as async reports is fully developed an in use.
Will need to consider redirects if new url structure as old route is
saved as user configuration.

### Fixes

- Pagination updated to not rely on the presence of a pathname which can
be undefined with nested routing
- Removed prefixes from docs as its currently unused and not working.