update establishment autocomplete with user context data (#458)
- Uses activeCaseloadId from the user context
- Checks for an input type of `autocomplete` that has `establishment` in
the name

this is not as robust as it could be with some BE work to provide a
unique key to identify the correct field to update with the user
context. However the current method of targeting the correct input will
work as a composite key as by convention the word `establishment` will
appear in the name of this field.