Tidy up mock reports (#464)
### Tidy up mock reports

This PR is to improve the content around the mock reports with the aim
to:
- remove mock report duplication
- make the content more readable
- Categorise the mock variants into relevantly named reports for
readability

This is part of a push to make the test-app a locally runnable app for
embedded teams to use as examples of how the lib should work, as well as
an internal tool we use for testing. In its current state it is
confusing and full of jargon that is not useful or readable.

### NOTE for reviewer.

No app code, or mock report code has actually changed so all files
changed are just configuration updates. They have just been moved to
relevant folders and their ID's have been updated. These ID's had to be
updated in the int and unit tests, and the paths updated.

---------

Co-authored-by: rossdowthwaite <4712683+rossdowthwaite@users.noreply.github.com>