Dpr2 2161 Test app routing refactor + automated accessibility testing (#459)
This PR is huge. For which I apologies greatly and I will explain. Its
not as bad as it looks at first glance. The majority is the routes
refactor which is simple routing boilerplate and config.

This PR is a refactor of the `test-app` to: 
- convert it from `js` to `ts`. 
- Use the same nested routing pattern that we use in the lib with a view
to make it easier to test. (this is where most of the file count comes
from)
- Add routes to all components, report templates & pages so they exist
in isolation
- Add cypress test to each testable path, so to move away from cucumber
- Automating the accessibility testing of all components, report
templates & pages so we have full automated accessibility coverage.
- Fix any accessibility issues

From high to low, the files changed are to do with:

- The routes refactor in the `test-app`. Just standard boilerplate
routing and page additions - controller.ts, routes.ts, tests.cy.ts
- Fix issue with nested routes in lib (see below)
- Updating tests to new paths.
- Fixing accessibility issues

### Fix issue with nested routes in lib

Converting the routes to nested routes revealed that the embedded
platform could not handle the nested routes and calls to apis and
platform routes broke. This would mean that teams would always have to
embed the platform at the '/' level and cant create their own path to
the report. This is fixed in this PR so teams can now nest the platform.

- Updated urls to recognise the nested route either by tweaking the url,
or prefixing the url with the nested route base
- Updated tests to use nested route.

### Next steps

Consolidate the existing cucumber tests into specific test files

---------

Co-authored-by: rossdowthwaite <4712683+rossdowthwaite@users.noreply.github.com>