Dpr download bug (#473)
Bug fix for reported bug in the in the ask_dpr slack channel - Report
was showing 11 columns but downloaded data only had 10

The column headings for download are collated from the first row of data
and referenced against the field definition to get the display name.
It was incorrectly assumed that all data rows would contain all column
header keys, even if the value was null/undefined. So in the case of the
bug, the first row only had 10 keys so missed off the 11th.