Dpr2 2400 zod dashboard validation (#560)
### Dashboard Visualisation validation
- Use Zod to validate vis definitions to ensure authors are providing
the correct definitions per visualisation type. Throws validation errors
when definition is incorrect for the specific type, to help guide the
author to creating a valid and working definition
- Refactored charts to be created the same way. Removed code duplication
and pulled client side processes and config in to the BE, so to lessen
the amount of client side JS code.
- Added new mocked dashboards to preview the dashboards.